+++ packages/egenie-utils/src/egGrid/egGridModel.tsx
@@ -105,6 +105,7 @@ export interface IEgGridModel {
   onSelectSum?: boolean;
   searchReduce?: boolean;
   searchReduceConfig?: TLabelName;
+  showGridOrderNo?: boolean;
 }
 
 export class EgGridModel {
@@ -331,6 +332,11 @@ export class EgGridModel {
    */
   @observable public searchReduceConfig: TLabelName = [];
 
+  /**
+   * 是否展示序号列
+   */
+  @observable public showGridOrderNo = true;
+
   @computed public get cacheKeyForColumnsConfig(): string {
     return `${this.user}_tsGrid_${ this.gridIdForColumnConfig}`;
   }
@@ -366,7 +372,7 @@ export class EgGridModel {
    * 组合序号列之后的列数据，渲染用，外部一般不用
    */
   @computed public get _columns() {
-    const { columns = [], showCheckBox = true, toggleOrDeleteSubRow, primaryKeyField } = this;
+    const { columns = [], showCheckBox = true, toggleOrDeleteSubRow, primaryKeyField, showGridOrderNo } = this;
     if (!columns.length) {
       return columns;
     }
@@ -420,7 +426,7 @@ export class EgGridModel {
         ...formatterCell,
       };
     });
-    const ret = (showCheckBox ? [SelectColumn] : []).concat([
+    const ret = (showCheckBox ? [SelectColumn] : []).concat(showGridOrderNo ? [
       {
         key: 'gridOrderNo',
         width: 50,
@@ -434,8 +440,8 @@ export class EgGridModel {
           </div>
         ),
       },
-      ...prevHandleColumns,
-    ]).filter((el: EnhanceColumn<IObj>) => !el.ejlHidden);
+    ] as ColumnType : []).concat([...prevHandleColumns])
+      .filter((el: EnhanceColumn<IObj>) => !el.ejlHidden);
 
     return ret;
   }
@@ -522,13 +528,13 @@ export class EgGridModel {
   public getFilterParams: () => {[key: string]: string; };
 
   public toggleOrDeleteSubRow = action(async({ id, type, primaryKeyField }: SubRowAction): Promise<void> => {
-    const { rows } = this;
+    const { rows, api: { onToggleOrDeleteSubRow }} = this;
 
-    if (type === 'toggleSubRow') {
+    if (type === 'toggleSubRow' && onToggleOrDeleteSubRow) {
       const rowIndex = rows.findIndex((r) => r[primaryKeyField] === id);
       const row = rows[rowIndex];
       if (!(row && row.isExpanded)) {
-        const reqRows = await this.api.onToggleOrDeleteSubRow?.({
+        const reqRows = await onToggleOrDeleteSubRow({
           id,
           type,
           primaryKeyField,
@@ -541,8 +547,8 @@ export class EgGridModel {
     }
 
     // 删除行且需要调用后端，且删除失败，那么什么都不做
-    if (type === 'deleteSubRow' && this.api.onToggleOrDeleteSubRow) {
-      const reqDeleteRow = await this.api.onToggleOrDeleteSubRow({
+    if (type === 'deleteSubRow' && onToggleOrDeleteSubRow) {
+      const reqDeleteRow = await onToggleOrDeleteSubRow({
         id,
         type,
         primaryKeyField,